+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+** Version 1.0.2 **
+* Clicking on a reputation bar now sets it as the tracked reputation.
+* Added a button to the default reputation frame to access BetterRep.
+
 ** Version 1.0.1 **
 * Removed 'The Mantid Swarm' faction from the Mists of Pandaria category, nobody likes them anyway.